+++ pybamm/models/submodels/electrolyte_conductivity/__init__.py
@@ -1,0 +1,6 @@
+from .base_electrolyte_conductivity import BaseElectrolyteConductivity
+from .leading_order_conductivity import LeadingOrder
+from .composite_conductivity import Composite
+from .full_conductivity import Full
+
+from . import surface_potential_form