+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# [Unreleased](https://github.com/pybamm-team/PyBaMM/)
+
+## Optimizations
+
+-   Made `QuickPlot` compatible with Google Colab ([#935](https://github.com/pybamm-team/PyBaMM/pull/935))
+-   Sped up model building ([#927](https://github.com/pybamm-team/PyBaMM/pull/927))
+-   Changed default solver for lead-acid to `CasadiSolver` ([#927](https://github.com/pybamm-team/PyBaMM/pull/927))
+
+## Bug fixes
+
+-   Reformatted electrolyte submodels ([#927](https://github.com/pybamm-team/PyBaMM/pull/927))
+
 # [v0.2.1](https://github.com/pybamm-team/PyBaMM/tree/v0.2.1) - 2020-03-31
 
 New expression tree node types, models, parameter sets and solvers, as well as general bug fixes and new examples.
@@ -11,9 +23,6 @@ New expression tree node types, models, parameter sets and solvers, as well as g
 -   Added functionality to broadcast to edges ([#891](https://github.com/pybamm-team/PyBaMM/pull/891))
 -   Reformatted and cleaned up `QuickPlot` ([#886](https://github.com/pybamm-team/PyBaMM/pull/886))
 -   Added thermal effects to lead-acid models ([#885](https://github.com/pybamm-team/PyBaMM/pull/885))
--   Add new symbols `VariableDot`, representing the derivative of a variable wrt time,
-    and `StateVectorDot`, representing the derivative of a state vector wrt time
-    ([#858](https://github.com/pybamm-team/PyBaMM/issues/858))
 -   Added a helper function for info on function parameters ([#881](https://github.com/pybamm-team/PyBaMM/pull/881))
 -   Added additional notebooks showing how to create and compare models ([#877](https://github.com/pybamm-team/PyBaMM/pull/877))
 -   Added `Minimum`, `Maximum` and `Sign` operators